Save @ 2021-07-30 17:03

+++ Configs/Position.cs
@@ -5,10 +5,10 @@ namespace Essentials.Configs
     public class Position
     {
         [JsonPropertyName("x")]
-        public double X { get; set; }
+        public float X { get; set; }
         [JsonPropertyName("y")]
-        public double Y { get; set; }
+        public float Y { get; set; }
         [JsonPropertyName("z")]
-        public double Z { get; set; }
+        public float Z { get; set; }
     }
 }
